Update EIP-6404: Add support for SetCode transactions
EIP-7702 adds SetCode transactions, adding another RLP format that now
needs conversion into SSZ. As part of this addition, the `ChainId` type
is changed to `uint256` to support EIP-7702 inconsistencies.

- Historically, `ChainId` was used in EIP-155 legacy transaction scheme
  as part of the signature's `v` value. There are no limits and no
  reserved values, and the math is assumed to not overflow.
- Practically, `ChainId` uses values that are safe to use on the web,
  limiting the space to `0 ... (2^53-1 - 35) / 2`. This motivated the
  earlier EIP-6404 design based on `uint64`.
- EIP-7702 allows using `1 ... 2^256-1` as `ChainId` for authorizations.

With `uint256` for `ChainId`, the same type supports both transaction
and authorization `ChainId`. Furthermore, `ChainId` 0 is reserved for
authorizations to indicate chain-agnostic authorizations. To reflect
that special meaning, the SSZ type uses `None` in authorizations.
This matches the behaviour for replayable transactions in SSZ.

Further, `SSZ_TX_TYPE` is changed to `0x1f` to avoid clashing with the
new envelope type `0x04` that is used for SetCode transactions.

- `0x05` is blocked because it's used as the signature domain for
  EIP-7702 authorizations
- `0x19` is blocked because it's used as the signature domain for
  general non-transaction messages
- Using a value above `0x19` avoids clashing with any RLP tx types.
- Using a value below `0x20` avoids giving bits `0x60` special meaning.

A commitment to `magic` is stored in the SSZ authorization object to
remember whether an authorization was originally made in RLP or a future
scheme.